chore(electron): Move off deprecated remote module
