+++ package.json
@@ -93,6 +93,7 @@
     "afterSign": "scripts/notarize.js"
   },
   "dependencies": {
+    "@electron/remote": "^2.0.11",
     "@primer/octicons-react": "19.8.0",
     "axios": "1.5.1",
     "date-fns": "2.30.0",
