+++ src/utils/auth.ts
@@ -1,7 +1,6 @@
 import { generateGitHubAPIUrl } from './helpers';
 
-const { remote } = require('electron');
-const BrowserWindow = remote.BrowserWindow;
+const { BrowserWindow } = require('@electron/remote');
 
 import { apiRequest, apiRequestAuth } from '../utils/api-requests';
 import { AuthResponse, AuthState, AuthTokenResponse } from '../types';
